@@ -18,72 +18,72 @@ apply plugin: 'com.github.dcendents.android-maven'
 apply plugin: 'com.jfrog.bintray'
 
 //////////////////////////////// Upload to BinTray ////////////////////////////////
-group = "com.azoft.carousellayoutmanager"
-final def siteUrl = 'https://github.com/Azoft/CarouselLayoutManager'
-final def gitUrl = "https://github.com/Azoft/CarouselLayoutManager.git"
-// This is the library version used when deploying the artifact
-version = "1.2.4"
-
-final Properties properties = new Properties()
-properties.load(project.rootProject.file('local.properties').newDataInputStream())
-
-bintray {
-    user = properties.getProperty("bintray.user")
-    key = properties.getProperty("bintray.apikey")
-
-    configurations = ["archives"]
-    pkg {
-        repo = "maven"
-        name = "carousel"
-        vcsUrl = gitUrl
-        licenses = ["Apache-2.0"]
-        publish = true
-    }
-}
-
-install {
-    repositories.mavenInstaller {
-        pom.project {
-            packaging 'jar'
-
-            // Add your description here
-            name "CarouselLayoutManager for RecyclerView in Android. Description is here: https://github.com/Azoft/CarouselLayoutManager"
-            url siteUrl
-
-            // Set your license
-            licenses {
-                license {
-                    name 'The Apache Software License, Version 2.0'
-                    url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
-                }
-            }
-            developers {
-                developer {
-                    id 'azoft'
-                    name 'azoft'
-                    email 'android-mobile@azoft.com'
-                }
-            }
-            scm {
-                connection gitUrl
-                developerConnection gitUrl
-                url siteUrl
-
-            }
-        }
-    }
-}
+//group = "com.azoft.carousellayoutmanager"
+//final def siteUrl = 'https://github.com/Azoft/CarouselLayoutManager'
+//final def gitUrl = "https://github.com/Azoft/CarouselLayoutManager.git"
+//// This is the library version used when deploying the artifact
+//version = "1.2.4"
+//
+//final Properties properties = new Properties()
+//properties.load(project.rootProject.file('local.properties').newDataInputStream())
+//
+//bintray {
+//    user = properties.getProperty("bintray.user")
+//    key = properties.getProperty("bintray.apikey")
+//
+//    configurations = ["archives"]
+//    pkg {
+//        repo = "maven"
+//        name = "carousel"
+//        vcsUrl = gitUrl
+//        licenses = ["Apache-2.0"]
+//        publish = true
+//    }
+//}
+//
+//install {
+//    repositories.mavenInstaller {
+//        pom.project {
+//            packaging 'jar'
+//
+//            // Add your description here
+//            name "CarouselLayoutManager for RecyclerView in Android. Description is here: https://github.com/Azoft/CarouselLayoutManager"
+//            url siteUrl
+//
+//            // Set your license
+//            licenses {
+//                license {
+//                    name 'The Apache Software License, Version 2.0'
+//                    url 'http://www.apache.org/licenses/LICENSE-2.0.txt'
+//                }
+//            }
+//            developers {
+//                developer {
+//                    id 'azoft'
+//                    name 'azoft'
+//                    email 'android-mobile@azoft.com'
+//                }
+//            }
+//            scm {
+//                connection gitUrl
+//                developerConnection gitUrl
+//                url siteUrl
+//
+//            }
+//        }
+//    }
+//}
 //////////////////////////////// Upload to BinTray ////////////////////////////////
 
 android {
-    compileSdkVersion 27
-    buildToolsVersion "27.0.3"
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion "28.0.3"
+    buildToolsVersion '28.0.3'
     resourcePrefix "carousellayoutmanager__"
 
     defaultConfig {
-        minSdkVersion 11
-        targetSdkVersion 27
+        minSdkVersion 14
+        targetSdkVersion 28
         versionCode 1
         versionName version
     }
