@@ -1,6 +1,6 @@
 package com.azoft.carousellayoutmanager;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 import android.view.View;
 
 /**
