@@ -1,6 +1,6 @@
 package com.azoft.carousellayoutmanager;
 
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 
 /**
  * Class for centering items after scroll event.<br />
